+++ CHANGELOG.md
@@ -43,20 +43,20 @@ All notable changes to the **FlexColorScheme** (FCS) package are documented here
 - Added theming of InputDecorator prefixIcon color.
 - Added theming of Drawer, NavigationDrawer and indicator width, as well as indicator border radius and indicator item color.
 - Added workaround for NavigationDrawer Flutter width and resize bugs. Bugs are still there, only worked around the cases to avoid them impacting the Playground.
+- Added theme simulator with **Themed Components**.
 
 
 **TODO BEFORE FCS STABLE 7.0 RELEASE**
 
 - REPORT: Found NavigationDrawer issues to Flutter repo.
 - FIX: Shortcut entry bug on MenuAnchor showcase.
-- INVESTIGATE: Playground NavigationBar seen to sometimes use wrong background default color label.  
 - ADD: Drawer indicator opacity.
 - ADD: Drawer indicator width to UI.
 - ADD: All navigation bars: Separate label and icon color in the Playground. API already supports it.
 - ADD: Better info about Android System Navigation Bar, plus link to docs, improve its docs.  
 - NEW: Bring shadows back in M3, maybe (Default, Adaptive, None, Custom).
 - NEW: Remove tint elevation in M3, maybe (Default, Adaptive, None, Custom).  
-- MAYBE ADD: TextFiled Demo setting example(s)
+- MAYBE ADD: TextFiled settings examples.
 - MAYBE ADD: Page with example theming configs.  
 - MAYBE ADD: Some initial theming options for the new MenuBar theme.
 - MAYBE REMOVE: ListTile transparent added to core defaults, may not be needed. This was temporarily needed in master, but should not be needed in stable 3.7, if not remove it before release.